Changed rand argument types from size_t to ul
Arguments to vmf::VmfRand::randBetween are now casted to (or declared as) unsigned long before being passed. The result is casted back to size_t to minimize codebase changes.

+++ Radamsa/vmf/src/modules/common/mutator/RadamsaDuplicateNodeMutator.cpp
@@ -142,9 +142,9 @@ void RadamsaDuplicateNodeMutator::mutateTestCase(StorageModule& storage, Storage
         return;
     }
 
-    const size_t lower{1u};
-    const size_t upper{numNodes - 2};
-    size_t nodeIndexToDuplicate{this->rand->randBetween(lower, upper)}; // not const, because findNodeByIndex will modify it
+    const unsigned long lower{1ul};
+    const unsigned long upper{static_cast<unsigned long>(numNodes - 2)};
+    size_t nodeIndexToDuplicate{static_cast<size_t>(this->rand->randBetween(lower, upper))}; // not const, because findNodeByIndex will modify it
     Node* nodeToDuplicate = tr.findNodeByIndex(tr.root, nodeIndexToDuplicate); 
 
     tr.duplicateNode(nodeToDuplicate, nodeToDuplicate->parent);
